% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countriesNBtoEN.R
\name{countries_nb_to_en}
\alias{countries_nb_to_en}
\title{Norwegian to English country names}
\usage{
countries_nb_to_en(save_loc = file.path("inst", "createddata"))
}
\description{
This function returns a named vector.
The names are Norwegian country names and the
values are English country names.
}
