% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days.r
\docType{data}
\name{days}
\alias{days}
\title{Dates of different days within isoweekyears}
\format{\describe{
\item{yrwk}{Isoweek-isoyear.}
\item{mon}{Date of Monday.}
\item{tue}{Date of Tuesday.}
\item{wed}{Date of Wednesday.}
\item{thu}{Date of Thursday.}
\item{fri}{Date of Friday.}
\item{sat}{Date of Saturday.}
\item{sun}{Date of Sunday.}
}}
\usage{
days
}
\description{
Dates of different days within isoweekyears
}
\keyword{datasets}
