% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLD_data_nordic.R
\docType{data}
\name{denmark_locations_long_b2020}
\alias{denmark_locations_long_b2020}
\title{Names of areas in Denmark that existed in 2020.}
\format{
\describe{
\item{iso3}{Location code.}
\item{granularity_geo}{nation, county, municip, ward}
\item{location_code}{Location code.}
\item{location_name}{Location name.}
}
}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
denmark_locations_long_b2020
}
\description{
Names of areas in Denmark that existed in 2020.
}
\keyword{datasets}
