% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordic.R
\docType{data}
\name{finland_population_by_age_b2020}
\alias{finland_population_by_age_b2020}
\title{Names of areas in Finland that existed in 2020.}
\format{
\describe{
\item{year}{Year}
\item{granularity_geo}{nation, county, municip, ward}
\item{location_code}{Location code.}
\item{age}{Age.}
\item{pop}{Population.}
}
}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
finland_population_by_age_b2020
}
\description{
Names of areas in Finland that existed in 2020.
}
\keyword{datasets}
