% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_countries_nb_to_en.R
\name{gen_countries_nb_to_en}
\alias{gen_countries_nb_to_en}
\title{Norwegian to English country names}
\usage{
gen_countries_nb_to_en()
}
\description{
This function returns a named vector.
The names are Norwegian country names and the
values are English country names.
}
