% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.r
\name{gen_data}
\alias{gen_data}
\title{Creates the norway_locations, norway_municip_merging, and norway_population data.table}
\usage{
gen_data(save_loc = file.path("inst", "createddata"))
}
\arguments{
\item{save_loc}{Location of data}
}
\description{
Creates the norway_locations, norway_municip_merging, and norway_population data.table
}
