% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_all.r
\name{gen_data_all}
\alias{gen_data_all}
\title{Creates the norway_locations, norway_municip_merging, and norway_population data.table}
\usage{
gen_data_all(base_loc = getwd())
}
\arguments{
\item{base_loc}{Location of data}
}
\description{
Creates the norway_locations, norway_municip_merging, and norway_population data.table
}
