% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.r
\name{gen_norway_locations}
\alias{gen_norway_locations}
\title{Creates the norway_locations data.table}
\usage{
gen_norway_locations(save_loc = file.path("inst", "createddata"))
}
\arguments{
\item{save_loc}{Location to save file to}
}
\description{
Creates the norway_locations data.table
}
