% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_municip_merging.r
\name{gen_norway_municip_merging}
\alias{gen_norway_municip_merging}
\title{Creates the norway_municip_merging (kommunesammenslaaing) data.table}
\usage{
gen_norway_municip_merging()
}
\description{
Last updated 2019-03-14
}
