% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_population.r
\name{gen_norway_population}
\alias{gen_norway_population}
\title{Creates the population dataset
https://www.ssb.no/en/statbank/table/07459/tableViewLayout1/}
\usage{
gen_norway_population(is_current_municips = TRUE)
}
\arguments{
\item{is_current_municips}{Do you want the population file to contain the current municipalities (i.e. after municipal merging) or the original municipalities that existed in that year?}
}
\description{
Creates the population dataset
https://www.ssb.no/en/statbank/table/07459/tableViewLayout1/
}
