% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.r
\name{get_data}
\alias{get_data}
\title{Get data.table sets from this package}
\usage{
get_data(name, is_current_municips = NULL, ...)
}
\arguments{
\item{name}{The name of the dataset. This interacts with \code{is_current_municips}.
If \code{is_current_municips} is \code{TRUE}/`FALSE``, then the following options are allowed:
\itemize{
\item norway_population: Population data
\item norway_locations: One row for each municipality containing columns 'municip_code', 'municip_name', 'county_code', and 'county_name'
\item norway_locations_long: One column for 'code', one column for 'name'
\item norway_municip_merging: Used to convert original municipalities into current municipalities
}

Some of these datasets (\code{norway_population}, \code{norway_locations}, and \code{norway_locations_long}) have the option to display the data using municipality codes as they were originally or as they are today (after the merging of municipalities).

If \code{is_current_municips} is `NULL``, then the following options are allowed:
\itemize{
\item norway_population_current
\item norway_population_original
\item norway_locations_current
\item norway_locations_original
\item norway_locations_long_current
\item norway_locations_long_original
\item norway_municip_merging
}}

\item{is_current_municips}{If this is \code{NULL}, then \code{name} is used exclusively. If this is \code{, then `name=name_current`, and if this is} then \code{name=name_original}. This lets you work more programatically.}

\item{...}{Not used currently}
}
\description{
Get data.table sets from this package
}
\examples{
get_data("norway_population", is_current_municips=TRUE)
get_data("norway_population_current")
}
