% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_icpc2_code.R
\docType{data}
\name{icpc2_code_en_2018}
\alias{icpc2_code_en_2018}
\title{ICPC-2: The International Classification of Primary Care}
\format{
\describe{
\item{Code}{ICPC2 code (2018)}
\item{shortTitle}{Description of the code.}
}
}
\source{
\url{https://ehelse.no/kodeverk/icpc-2e--english-version}
}
\usage{
icpc2_code_en_2018
}
\description{
ICPC-2: The International Classification of Primary Care
}
\keyword{datasets}
