% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_icpc2_code.R
\docType{data}
\name{icpc2_code_nb_2020}
\alias{icpc2_code_nb_2020}
\title{ICPC-2: The International Classification of Primary Care (Norwegian)}
\format{
\describe{
\item{Code}{ICPC2 code (2020)}
\item{shortTitle_nb}{Description of the code in Norwegian.}
}
}
\source{
\url{https://ehelse.no/kodeverk/icpc-2.den-internasjonale-klassifikasjonen-for-primaerhelsetjenesten}
}
\usage{
icpc2_code_nb_2020
}
\description{
ICPC-2: The International Classification of Primary Care (Norwegian)
}
\keyword{datasets}
