% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{make_skeleton}
\alias{make_skeleton}
\title{make_skeleton}
\usage{
make_skeleton(
  date_min = NULL,
  date_max = NULL,
  isoyearweek_min = NULL,
  isoyearweek_max = NULL,
  time_total = FALSE,
  location_code = NULL,
  granularity_geo = "all",
  location_reference = fhidata::norway_locations_names(),
  ...
)
}
\arguments{
\item{date_min}{The minimum date for the skeleton}

\item{date_max}{The maximum date for the skeleton}

\item{isoyearweek_min}{The minimum isoyearweek for the skeleton}

\item{isoyearweek_max}{The maximum isoyearweek for the skeleton}

\item{time_total}{Producing a 'total' time for the skeleton}

\item{location_code}{The location_code's wanted for the skeleton}

\item{granularity_geo}{The granularity_geo's wanted for the skeleton. If this is a list, then the returned value will also be in a list (see examples).}

\item{location_reference}{A data.table that contains two columns: location_code and granularity_geo}

\item{...}{Other variables to include in the skeleton}
}
\description{
The easiest way to make structural data skeletons.
}
\examples{
make_skeleton(date_min="2020-01-01", date_max="2020-01-30", granularity_geo = c("nation", "county"))[]
make_skeleton(isoyearweek_min="2020-01", isoyearweek_max="2020-15", granularity_geo = c("nation", "county"))[]
make_skeleton(time_total = TRUE, granularity_geo = c("nation", "wardoslo"))[]
make_skeleton(time_total = TRUE, granularity_geo = list(c("wardoslo", "missingwardoslo"), c("county", "missingcounty", "notmainlandcounty"), "nation"))[]
make_skeleton(time_total = TRUE, granularity_geo = list("firstbatch" = c("wardoslo", "missingwardoslo"), c("county", "missingcounty", "notmainlandcounty"), "nation"))[]
make_skeleton(isoyearweek_min = "2020-01", isoyearweek_max = "2020-03", time_total = TRUE, granularity_geo = c("nation", "wardoslo"))[]
}
