% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_environments.r
\docType{data}
\name{nor_char}
\alias{nor_char}
\title{Norwegian characters in unicode}
\format{An object of class \code{environment} of length 6.}
\usage{
nor_char
}
\description{
Norwegian characters in unicode
}
\keyword{datasets}
