% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_childhood_vax.r
\docType{data}
\name{norway_childhood_vax}
\alias{norway_childhood_vax}
\title{Childhood vaccination rates in Norway}
\format{\describe{
\item{year}{The middle year of a 5 year range (e.g. 2011 is the average of data from 2009-2013).}
\item{location_code}{The location code.}
\item{age}{The population age.}
\item{vax}{The vaccine.}
\item{proportion}{Proportion of people who are vaccinated.}
\item{imputed}{FALSE if real data. TRUE if it is the national average.}
}}
\source{
\url{http://khs.fhi.no/webview/}
}
\usage{
norway_childhood_vax
}
\description{
We conveniently package vaccine coverage data taken from "Kommunehelsa statistikkbank".
This data was last updated on 2019-04-09.
}
\details{
This dataset contains national/county/municipality level (5 year average) vaccination coverage rates
for 16 year olds for the childhood vaccination program (diphtheria, hpv, measles,
mumps, poliomyelitis, pertussis, rubella, tetanus).

Municipalities are updated for the 2019 redistricting.
}
\keyword{datasets}
