% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_municip_merging.r
\docType{data}
\name{norway_county_merging_b2019}
\alias{norway_county_merging_b2019}
\title{County redistricting Norway (2019 borders).}
\format{\describe{
\item{county_code_current}{The county code per today.}
\item{county_code_original}{The county code as of 'year'.}
\item{year}{The year corresponding to 'county_code_original'.}
\item{weighting}{The weighting that needs to be applied.}
}}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
norway_county_merging_b2019
}
\description{
This dataset is used to transfer "original" municipality level datasets to the 2019 borders.
}
\details{
Last updated 2019-03-14
}
\keyword{datasets}
