% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dates_holidays.r
\docType{data}
\name{norway_dates_holidays}
\alias{norway_dates_holidays}
\title{Dates of Norwegian holidays}
\format{\describe{
\item{date}{Date.}
\item{is_holiday}{Is it a holiday?}
}}
\usage{
norway_dates_holidays
}
\description{
Dates of Norwegian holidays
}
\keyword{datasets}
