% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_municip_merging.r
\docType{data}
\name{norway_fixing_merged_municips}
\alias{norway_fixing_merged_municips}
\title{Fixing data that has already been redistricted.}
\format{\describe{
\item{municip_code_current}{The municipality code per today.}
\item{municip_code_original}{The municipality code as of 'year'.}
\item{year}{The year corresponding to 'municip_code_original'.}
\item{border_start}{The year that the data has currently been redistricted to.}
\item{border_end}{The year of the desired final redistricting.}
\item{weighting}{The weighting that needs to be applied.}
}}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
norway_fixing_merged_municips
}
\description{
This dataset is used to transfer municipality level datasets (that have already
been redistricted) to future borders. If `border_start=2018` then it is
assumed that the 2018 borders were used from 2000-2018, and then 2019 borders
used for 2019, 2020 borders for 2020, and so on. Hence a combination of `year`
and `border_start` must always be used.
}
\details{
Last updated 2019-03-14
}
\keyword{datasets}
