% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_locations.r
\docType{data}
\name{norway_locations_b2019}
\alias{norway_locations_b2019}
\title{Names of areas in Norway that existed in 2019.}
\format{\describe{
\item{municip_code}{Municipality code.}
\item{municip_name}{Municipality name.}
\item{county_code}{County code.}
\item{county_name}{County name.}
}}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
norway_locations_b2019
}
\description{
Names of areas in Norway that existed in 2019.
}
\keyword{datasets}
