% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_locations.r
\docType{data}
\name{norway_locations_current}
\alias{norway_locations_current}
\title{Names of areas in Norway that currently exist.}
\format{\describe{
\item{municip_code}{Municipality code.}
\item{municip_name}{Municipality name.}
\item{county_code}{County code.}
\item{county_name}{County name.}
}}
\source{
\url{https://snl.no/kommunenummer}
}
\usage{
norway_locations_current
}
\description{
Names of areas in Norway that currently exist.
}
\keyword{datasets}
