% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_locations_long.r
\docType{data}
\name{norway_locations_long_current}
\alias{norway_locations_long_current}
\title{Names of areas in Norway that currently exist.}
\format{\describe{
\item{location_code}{Location code.}
\item{location_name}{Location name.}
}}
\source{
\url{https://snl.no/kommunenummer}
}
\usage{
norway_locations_long_current
}
\description{
Names of areas in Norway that currently exist.
}
\keyword{datasets}
