% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_locations_long.r
\docType{data}
\name{norway_locations_long_original}
\alias{norway_locations_long_original}
\title{Names of areas in Norway that previously existed.}
\format{\describe{
\item{location_code}{Location code.}
\item{location_name}{Location name.}
}}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
norway_locations_long_original
}
\description{
Names of areas in Norway that previously existed.
}
\keyword{datasets}
