% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norway_locations_msis_to_fhidata.R
\name{norway_locations_msis_to_fhidata}
\alias{norway_locations_msis_to_fhidata}
\title{Converts MSIS' location values to fhidata location_codes}
\usage{
norway_locations_msis_to_fhidata(x)
}
\arguments{
\item{x}{A vector of numbers that correspond to MSIS' location values}
}
\description{
Converts MSIS' location values to fhidata location_codes
}
\examples{
norway_locations_msis_to_fhidata(c(0, 3, 21, 50, 99, 9999, 301, 5001, 9999, 30101, 30116, 30199, NA))
}
