% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLD_data_norway_locations.r
\docType{data}
\name{norway_locations_municip_b2020}
\alias{norway_locations_municip_b2020}
\title{Names of municipality in Norway that existed in 2020.}
\format{
\describe{
\item{municip_code}{Municipality code.}
\item{municip_name}{Municipality name.}
\item{county_code}{County code.}
\item{county_name}{County name.}
\item{region_code}{Region code.}
\item{region_name}{Region name.}
\item{faregion_code}{Food authority region code.}
\item{faregion_name}{Food authority region name.}
\item{baregion_code}{Bo- og arbeids region code.}
\item{baregion_name}{Bo- og arbeids region name.}
}
}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
norway_locations_municip_b2020
}
\description{
(same as norway_locations_b2020)
}
\keyword{datasets}
