% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norway_locations_names.R
\name{norway_locations_names}
\alias{norway_locations_names}
\title{All names in Norway (programmable borders).}
\usage{
norway_locations_names(border = fhidata::config$border)
}
\arguments{
\item{border}{The border year}
}
\description{
All names in Norway (programmable borders).
}
\examples{
norway_locations_names()
}
