% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norway_locations_names.R
\docType{data}
\name{norway_locations_names_b2020}
\alias{norway_locations_names_b2020}
\title{Names of areas in Norway that existed in 2020.}
\format{
\describe{
\item{location_code}{Location code.}
\item{location_name}{Location name.}
\item{location_name_description}{Location name with additional description.}
\item{location_order}{The preferred presentation order}
\item{granularity_geo}{nation, county, municip, wardoslo, wardbergen, wardstavanger, wardtrondheim, baregion}
}
}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
norway_locations_names_b2020
}
\description{
Names of areas in Norway that existed in 2020.
}
\keyword{datasets}
