% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_locations.r
\docType{data}
\name{norway_locations_notmainland_b2020}
\alias{norway_locations_notmainland_b2020}
\title{Names of areas in Norway (not mainland, i.e. Svalbard).}
\format{
\describe{
\item{municip_code}{Municipality code.}
\item{municip_name}{Municipality name.}
\item{county_code}{County code.}
\item{county_name}{County name.}
\item{region_code}{Region code.}
\item{region_name}{Region name.}
\item{faregion_code}{Food authority region code.}
\item{faregion_name}{Food authority region name.}
}
}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
norway_locations_notmainland_b2020
}
\description{
Names of areas in Norway (not mainland, i.e. Svalbard).
}
\keyword{datasets}
