% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norway_locations_redistricting_all_b2020.R
\name{norway_locations_redistricting}
\alias{norway_locations_redistricting}
\title{All redistricting in Norway (programable borders).}
\usage{
norway_locations_redistricting(border = fhidata::config$border)
}
\arguments{
\item{border}{The border year}
}
\description{
All redistricting in Norway (programable borders).
}
\examples{
norway_locations_redistricting()
}
