% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_locations.r
\docType{data}
\name{norway_locations_ward_b2020}
\alias{norway_locations_ward_b2020}
\title{Names of wards (bydel, city district) in Norway that existed in 2020.}
\format{
\describe{
\item{ward_code}{Ward (bydel) code.}
\item{ward_name}{Ward (bydel) name.}
\item{municip_code}{Municipality code.}
\item{municip_name}{Municipality name.}
}
}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
norway_locations_ward_b2020
}
\description{
Names of wards (bydel, city district) in Norway that existed in 2020.
}
\keyword{datasets}
