% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_map.R
\name{norway_map_counties}
\alias{norway_map_counties}
\title{norway_map_counties}
\usage{
norway_map_counties(save_loc = file.path("inst", "createddata"))
}
\arguments{
\item{save_loc}{Location to save file to}
}
\description{
norway_map_counties
}
