% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_map.R
\docType{data}
\name{norway_map_counties_label_positions_b2019}
\alias{norway_map_counties_label_positions_b2019}
\title{Label Positions for Norwegian Counties (2019 borders)}
\format{\describe{
\item{location_code}{Location code (county code).}
\item{long}{Location code.}
\item{lat}{Location name.}
}}
\usage{
norway_map_counties_label_positions_b2019
}
\description{
GPS coordinates for labels for Norwegian counties
}
\examples{
library(ggplot2)
q <- ggplot(mapping = aes(x = long, y = lat))
q <- q + geom_polygon(
  data = fhidata::norway_map_counties_b2019,
  mapping = aes(group = group),
  color = "black",
  fill = "white",
  size = 0.2
)
q <- q + geom_text(
  data = fhidata::norway_map_counties_label_positions_b2019,
  mapping = aes(label = location_code),
  color = "red"
)
q <- q + theme_void()
q <- q + coord_quickmap()
q
}
\keyword{datasets}
