% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_map.R
\docType{data}
\name{norway_map_insert_title_position_b2020}
\alias{norway_map_insert_title_position_b2020}
\title{Title Positions for Inserts (2020 borders)}
\format{\describe{
\item{long}{Location code.}
\item{lat}{Location name.}
}}
\usage{
norway_map_insert_title_position_b2020
}
\description{
GPS coordinates for inserts for Norwegian maps
}
\examples{
library(ggplot2)
pd <- fhidata::norway_map_counties_with_insert_b2020
q <- ggplot()
q <- q + geom_polygon(
  data = pd,
  aes(x = long, y = lat, group = group, fill = location_code),
  color = "black"
)
q <- q + annotate(
  "text",
  x = fhidata::norway_map_insert_title_position_b2020$long,
  y = fhidata::norway_map_insert_title_position_b2020$lat,
  label = "Oslo"
)
q <- q + theme_void()
q <- q + coord_quickmap()
q
}
\keyword{datasets}
