% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_map.R
\name{norway_map_municips}
\alias{norway_map_municips}
\title{norway_map_municips
https://kartkatalog.geonorge.no/metadata/uuid/cb02ab77-d3e6-4500-8a92-ea67367e7734}
\usage{
norway_map_municips(save_loc = file.path("inst", "createddata"))
}
\arguments{
\item{save_loc}{Location to save file to}
}
\description{
norway_map_municips
https://kartkatalog.geonorge.no/metadata/uuid/cb02ab77-d3e6-4500-8a92-ea67367e7734
}
