% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_map.R
\docType{data}
\name{norway_map_municips_b2020}
\alias{norway_map_municips_b2020}
\title{Maps of Norwegian Municipalities (2020 borders)}
\format{\describe{
\item{long}{Location code.}
\item{lat}{Location name.}
\item{order}{The order that this line should be plotted in.}
\item{hole}{Geodata. Not used.}
\item{piece}{Geodata. Not used.}
\item{group}{Needs to be used as 'group' aesthetic in ggplot2.}
\item{id}{Non-informative id code.}
\item{location_code}{Location code (municipality code).}
}}
\source{
\url{https://kartkatalog.geonorge.no/metadata/f015cd24-a14b-42f3-b695-848cb88d9657}
}
\usage{
norway_map_municips_b2020
}
\description{
We conveniently package map datasets for Norwegian municipalities
(taken from Geonorge) that can be used in ggplot2 without needing any geo
libraries. This data is licensed under Creative Commons BY 4.0 (CC BY 4.0).
}
\examples{
library(ggplot2)
q <- ggplot(mapping = aes(x = long, y = lat, group = group))
q <- q + geom_polygon(data = fhidata::norway_map_municips_b2020, color = "black")
q <- q + theme_void()
q <- q + coord_quickmap()
q
}
\keyword{datasets}
