% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLD_data_norway_municip_merging_missing.r
\docType{data}
\name{norway_missingmunicip_merging_b2020}
\alias{norway_missingmunicip_merging_b2020}
\title{Municipality redistricting in Norway (2020 borders).}
\format{
\describe{
\item{municip_code_current}{The municipality code per today.}
\item{municip_code_original}{The municipality code as of 'year'.}
\item{year}{The year corresponding to 'municip_code_original'.}
\item{weighting}{The weighting that needs to be applied.}
}
}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
norway_missingmunicip_merging_b2020
}
\description{
This dataset is used to transfer "original" municipality level datasets to the 2020 borders.
}
\details{
Last updated 2019-11-24
}
\keyword{datasets}
