% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_municip_merging.r
\docType{data}
\name{norway_municip_merging}
\alias{norway_municip_merging}
\title{Redistricting in Norway.}
\format{\describe{
\item{municip_code_current}{The municipality code per today.}
\item{municip_code_original}{The municipality code as of 'year'.}
\item{year}{The year corresponding to 'municip_code_original'.}
\item{municip_name}{The municipality name per today.}
\item{county_code}{The county code per today.}
\item{county_name}{The county name per today.}
\item{region_code}{The region code per today.}
\item{region_name}{The region name per today.}
}}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
norway_municip_merging
}
\description{
This dataset is used to transfer "original" municipality level datasets to the current redistricting.
}
\details{
Last updated 2019-03-14
}
\keyword{datasets}
