% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_municip_merging.r
\docType{data}
\name{norway_municip_merging_b2020}
\alias{norway_municip_merging_b2020}
\title{Municipality redistricting in Norway (2020 borders).}
\format{\describe{
\item{municip_code_current}{The municipality code per today.}
\item{municip_code_original}{The municipality code as of 'year'.}
\item{year}{The year corresponding to 'municip_code_original'.}
\item{municip_name}{The municipality name per today.}
\item{county_code}{The county code per today.}
\item{county_name}{The county name per today.}
\item{region_code}{The region code per today.}
\item{region_name}{The region name per today.}
\item{weighting}{The weighting that needs to be applied.}
}}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
norway_municip_merging_b2020
}
\description{
This dataset is used to transfer "original" municipality level datasets to the 2020 borders.
}
\details{
Last updated 2019-11-24
}
\keyword{datasets}
