% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLD_data_norway_municip_merging.r
\docType{data}
\name{norway_ward_merging_b2020}
\alias{norway_ward_merging_b2020}
\title{Ward (bydel) redistricting in Norway (2020 borders).}
\format{
\describe{
\item{ward_code_current}{The ward code per today.}
\item{ward_code_original}{The ward code as of 'year'.}
\item{year}{The year corresponding to 'ward_code_original'.}
\item{weighting}{The weighting that needs to be applied.}
}
}
\source{
\url{https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer}
}
\usage{
norway_ward_merging_b2020
}
\description{
This dataset is used to transfer "original" ward level datasets to the 2020 borders.
}
\details{
Last updated 2020-06-08
}
\keyword{datasets}
