% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frost.R
\docType{data}
\name{norway_weather_stations_max_air_temperature}
\alias{norway_weather_stations_max_air_temperature}
\title{Weather station list for max air temperatures in Norway}
\format{\describe{
\item{location_code}{The location code.}
\item{source}{Station code.}
}}
\source{
\url{https://frost.met.no/}
}
\usage{
norway_weather_stations_max_air_temperature
}
\description{
We create a list of weather stations that contain max air temperature data for
each municipality in Norway. When a municipality does not have a relevant weather
stations, we choose a weather station that is close to the municipality.
}
\details{
Municipalities are updated for the 2019 redistricting.
}
\keyword{datasets}
