% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_world_map.R
\docType{data}
\name{world_map}
\alias{world_map}
\title{World Map}
\format{\describe{
\item{long}{Location code.}
\item{lat}{Location name.}
\item{order}{The order that this line should be plotted in.}
\item{group}{Needs to be used as 'group' aesthetic in ggplot2.}
\item{location_code}{3 character isocode.}
}}
\source{
\url{https://www.naturalearthdata.com/downloads/50m-cultural-vectors/}
}
\usage{
world_map
}
\description{
We conveniently package map datasets for the world
(taken from Natural Earth) that can be used in ggplot2 without needing any geo
libraries. This data is licensed under the public domain.
}
\details{
This world map contains 196 administrative regions, corresponding to: 193 UN member states,
2 non-member permanent observer states (Holy See and the State of Palestine), and Kashmir (KAH).
}
\examples{
library(ggplot2)
q <- ggplot(mapping = aes(x = long, y = lat, group = group, fill = location_code))
q <- q + geom_polygon(data = fhidata::world_map, color = "black")
q <- q + theme_void()
q <- q + coord_quickmap()
q
}
\keyword{datasets}
