% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_countries_nb_to_en.R
\docType{data}
\name{countries_nb_to_en}
\alias{countries_nb_to_en}
\title{Country names}
\format{\describe{
\item{nb}{Country name in Norwegian bokmal.}
\item{en}{Country name in English.}
}}
\source{
\url{http://stat-computing.org/dataexpo/2006/}
}
\usage{
countries_nb_to_en
}
\description{
Country names
}
\keyword{datasets}
