% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_locations.r
\name{gen_norway_locations}
\alias{gen_norway_locations}
\title{Creates the norway_locations data.table}
\usage{
gen_norway_locations(is_current_municips = TRUE)
}
\arguments{
\item{is_current_municips}{Do you want the population file to contain the current municipalities (i.e. after municipal merging) or the original municipalities that existed in that year?}
}
\description{
Creates the norway_locations data.table
}
