% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.r
\name{gen_norway_locations_long}
\alias{gen_norway_locations_long}
\title{Creates the norway_locations data.table}
\usage{
gen_norway_locations_long(is_current_municips = TRUE,
  save_loc = file.path("inst", "createddata"))
}
\arguments{
\item{is_current_municips}{Do you want the population file to contain the current municipalities (i.e. after municipal merging) or the original municipalities that existed in that year?}

\item{save_loc}{Location to save file to}
}
\description{
Creates the norway_locations data.table
}
