% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norway_locations_hierarchy.R
\name{norway_locations_hierarchy}
\alias{norway_locations_hierarchy}
\title{Hierarchies in Norway (programable borders).}
\usage{
norway_locations_hierarchy(from, to, border = fhidata::config$border)
}
\arguments{
\item{from}{wardoslo, wardbergen, wardtrondheim, wardstavanger, municip, baregion, county, region, faregion, notmainlandmunicip, notmainlandcounty, missingmunicip, missingcounty}

\item{to}{wardoslo, wardbergen, wardtrondheim, wardstavanger, municip, baregion, county, region, faregion, notmainlandmunicip, notmainlandcounty, missingmunicip, missingcounty}

\item{border}{The border year}
}
\description{
Hierarchies in Norway (programable borders).
}
\examples{
norway_locations_hierarchy(from="wardoslo", to="county")
norway_locations_hierarchy(from="municip", to="baregion")
}
