% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norway_population_by_age_sex.R
\name{norway_population_by_age_sex_cats}
\alias{norway_population_by_age_sex_cats}
\title{norway_population_by_age_sex_cats}
\usage{
norway_population_by_age_sex_cats(cats = NULL, border = fhidata::config$border)
}
\arguments{
\item{cats}{A list containing vectors that you want to categorize}

\item{border}{The border year}
}
\description{
A function that easily categorizes the populations for you
}
\examples{
norway_population_by_age_sex_cats(cats = list(c(1:10), c(11:20)))
norway_population_by_age_sex_cats(cats = list("one to ten" = c(1:10), "eleven to twenty" = c(11:20)))
norway_population_by_age_sex_cats(cats = list(c(1:10), c(11:20), "21+"=c(21:200)))
}
